+++ lib/flutter_appbar.dart
@@ -9,5 +9,7 @@ export "components/nested_scroll_activity.dart";
 export "components/nested_scroll_controller.dart";
 
 export "widgets/appbar_connection.dart";
+export "widgets/appbar_column.dart";
+export "widgets/appbar_fade_effect.dart";
 export "widgets/nested_scroll_connection.dart";
 export "widgets/scrollable_gesture_delegator.dart";